requirements.txt
docopt
libvirt-python
py2neo
python-see
seaborn
